requirements.txt
# default docusign requirements
astroid==3.3.5
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
Click
cryptography==43.0.3
docusign-esign
docusign-rooms==1.3.0
docusign-monitor==1.2.0
docusign-click==1.4.0
docusign-admin==2.0.0rc2
docusign-webforms==1.0.2rc12
Flask==2.3.3
Flask-OAuthlib==0.9.6
Flask-Session==0.8.0
flask-wtf==1.2.1
flake8==7.1.1
idna==3.10
isort==5.13.2
itsdangerous==2.2.0
Jinja2>=3.1.4
lazy-object-proxy==1.10.0
MarkupSafe==2.1.5
mccabe==0.7.0
oauthlib==2.1.0
pipenv==2024.2.0
py-oauth2==0.0.10
pycparser==2.22
pylint==3.3.1
python-dateutil==2.9.0
python-dotenv==1.0.1
requests>=2.32.3
requests_auth>=8.0.0
requests-oauthlib==1.1.0
six==1.16.0
urllib3>=2.2.3
virtualenv==20.27.0
virtualenv-clone==0.5.7
Werkzeug==2.3.8
wrapt==1.16.0

# project requirements
streamlit
streamlit-oauth
streamlit-navigation-bar